Fix fragment doesn't save state

+++ app/src/main/java/com/example/sallachallenge/models/details/Rating.kt
@@ -5,5 +5,5 @@ data class Rating (
 
 	val total : Int,
 	val count : Int,
-	val rate : Int
+	val rate : Double
 )
